lntest/test: move multi-hop payments test into own file
